+++ lib/widgets/interactive/text.dart
