+++ lib/widgets/interactive/text_span.dart
